# doc/api/openapi.yaml
"openapi": "3.1.0"
"info":
  "title": "PixelPast"
  "version": "0.1.0"
"paths":
...
  "/api/exploration/bootstrap":
    "get":
      "tags":
        - "timeline"
      "summary": "Get exploration bootstrap metadata"
      "description": "Return exploration shell metadata for the resolved date range, including available view modes, known people, and tags, without the dense day grid payload."
      "operationId": "get_exploration_bootstrap_api_exploration_bootstrap_get"
      "parameters":
        -
          "name": "start"
          "in": "query"
          "required": false
          "schema":
            "anyOf":
              -
                "type": "string"
                "format": "date"
              -
                "type": "null"
            "description": "Inclusive UTC start date for the requested exploration window. Must be provided together with end."
            "title": "Start"
          "description": "Inclusive UTC start date for the requested exploration window. Must be provided together with end."
          "examples":
            "winter_window":
              "summary": "Window start for January 2026"
              "value": "2026-01-01"
        -
          "name": "end"
          "in": "query"
          "required": false
          "schema":
            "anyOf":
              -
                "type": "string"
                "format": "date"
              -
                "type": "null"
            "description": "Inclusive UTC end date for the requested exploration window. Must be provided together with start."
            "title": "End"
          "description": "Inclusive UTC end date for the requested exploration window. Must be provided together with start."
          "examples":
            "winter_window":
              "summary": "Window end for January 2026"
              "value": "2026-01-31"
      "responses":
        "200":
          "description": "Exploration shell metadata for the resolved inclusive UTC date range."
          "content":
            "application/json":
              "schema":
                "$ref": "#/components/schemas/ExplorationBootstrapResponse"
              "examples":
                "winter_overview":
                  "summary": "Bootstrap metadata for a winter exploration window"
                  "value":
                    "range":
                      "start": "2026-01-01"
                      "end": "2026-01-31"
                    "view_modes":
                      -
                        "id": "activity"
                        "label": "Activity"
                        "description": "Overall day activity score across all imported sources."
                      -
                        "id": "workdays_vacation"
                        "label": "Vacation"
                        "description": "Workday and vacation classification for each calendar day."
                    "persons":
                      -
                        "id": 7
                        "name": "Anna Becker"
                        "role": "family"
                      -
                        "id": 12
                        "name": "Milo Tan"
                        "role": "friend"
                    "tags":
                      -
                        "path": "travel/italy/venice"
                        "label": "Venice"
                      -
                        "path": "music/concert/live"
                        "label": "Live Concert"
        "400":
          "description": "The request was syntactically valid but violated a documented PixelPast API contract constraint."
          "content":
            "application/json":
              "examples":
                "invalid_range":
                  "summary": "Start date after end date"
                  "value":
                    "detail": "start must be less than or equal to end"
                "partial_range":
                  "summary": "Only one range boundary provided"
                  "value":
                    "detail": "start and end must both be provided together"
              "schema":
                "$ref": "#/components/schemas/ApiErrorResponse"
        "422":
          "description": "The request could not be parsed or failed declared parameter validation."
          "content":
            "application/json":
              "examples":
                "invalid_date":
                  "summary": "Unparseable ISO date"
                  "value":
                    "detail":
                      -
                        "type": "date_from_datetime_parsing"
                        "loc":
                          - "query"
                          - "start"
                        "msg": "Input should be a valid date or datetime, input is too short"
                        "input": "2026-02"
                "invalid_path_date":
                  "summary": "Invalid path day format"
                  "value":
                    "detail":
                      -
                        "type": "date_from_datetime_parsing"
                        "loc":
                          - "path"
                          - "day"
                        "msg": "Input should be a valid date or datetime, invalid date separator"
                        "input": "2026/07/14"
              "schema":
                "$ref": "#/components/schemas/ApiValidationErrorResponse"
...
  "/api/days/{day}":
    "get":
      "tags":
        - "timeline"
      "summary": "Get one day timeline"
      "description": "Return the unified canonical event and asset timeline for a single UTC calendar day, ordered by timestamp."
      "operationId": "get_day_detail_api_days__day__get"
      "parameters":
        -
          "name": "day"
          "in": "path"
          "required": true
          "schema":
            "type": "string"
            "format": "date"
            "description": "UTC calendar day to inspect in ISO format (YYYY-MM-DD)."
            "title": "Day"
          "description": "UTC calendar day to inspect in ISO format (YYYY-MM-DD)."
          "examples":
            "summer_day":
              "summary": "One summer travel day"
              "value": "2026-07-14"
      "responses":
        "200":
          "description": "Chronological event and asset items for one UTC day."
          "content":
            "application/json":
              "schema":
                "$ref": "#/components/schemas/DayDetailResponse"
              "examples":
                "mixed_day_timeline":
                  "summary": "Chronological event and asset timeline for one day"
                  "value":
                    "date": "2026-07-14"
                    "items":
                      -
                        "item_type": "event"
                        "id": 1042
                        "timestamp": "2026-07-14T08:30:00Z"
                        "event_type": "calendar"
                        "title": "Flight to Lisbon"
                        "summary": "TXL to LIS with morning check-in"
                      -
                        "item_type": "asset"
                        "id": 8801
                        "timestamp": "2026-07-14T14:12:43Z"
                        "media_type": "image"
                        "external_id": "photos/2026/07/IMG_1042.JPG"
                      -
                        "item_type": "event"
                        "id": 1043
                        "timestamp": "2026-07-14T19:00:00Z"
                        "event_type": "music_play"
                        "title": "Fado Playlist"
                        "summary": "Spotify playback during evening walk"
        "422":
          "description": "The request could not be parsed or failed declared parameter validation."
          "content":
            "application/json":
              "examples":
                "invalid_date":
                  "summary": "Unparseable ISO date"
                  "value":
                    "detail":
                      -
                        "type": "date_from_datetime_parsing"
                        "loc":
                          - "query"
                          - "start"
                        "msg": "Input should be a valid date or datetime, input is too short"
                        "input": "2026-02"
                "invalid_path_date":
                  "summary": "Invalid path day format"
                  "value":
                    "detail":
                      -
                        "type": "date_from_datetime_parsing"
                        "loc":
                          - "path"
                          - "day"
                        "msg": "Input should be a valid date or datetime, invalid date separator"
                        "input": "2026/07/14"
              "schema":
                "$ref": "#/components/schemas/ApiValidationErrorResponse"
...
"components":
  "schemas":
    "ApiErrorResponse":
      "properties":
        "detail":
          "type": "string"
          "title": "Detail"
      "type": "object"
      "required":
        - "detail"
      "title": "ApiErrorResponse"
      "description": "Error payload for contract-level request rejections."
    "ApiValidationErrorItem":
      "properties":
        "type":
          "type": "string"
          "title": "Type"
        "loc":
          "items":
            "anyOf":
              -
                "type": "string"
              -
                "type": "integer"
          "type": "array"
          "title": "Loc"
        "msg":
          "type": "string"
          "title": "Msg"
        "input":
          "anyOf":
            - {}
            -
              "type": "null"
          "title": "Input"
      "type": "object"
      "required":
        - "type"
        - "loc"
        - "msg"
      "title": "ApiValidationErrorItem"
      "description": "One FastAPI request validation error entry."
    "ApiValidationErrorResponse":
      "properties":
        "detail":
          "items":
            "$ref": "#/components/schemas/ApiValidationErrorItem"
          "type": "array"
          "title": "Detail"
      "type": "object"
      "required":
        - "detail"
      "title": "ApiValidationErrorResponse"
      "description": "Error payload returned when declared request validation fails."
    "DayAssetItem":
      "properties":
        "item_type":
          "type": "string"
          "const": "asset"
          "title": "Item Type"
        "id":
          "type": "integer"
          "title": "Id"
        "timestamp":
          "type": "string"
          "format": "date-time"
          "title": "Timestamp"
        "media_type":
          "type": "string"
          "title": "Media Type"
        "external_id":
          "type": "string"
          "title": "External Id"
      "type": "object"
      "required":
        - "item_type"
        - "id"
        - "timestamp"
        - "media_type"
        - "external_id"
      "title": "DayAssetItem"
      "description": "Serialized day-detail item for a canonical asset."
...
    "DayDetailResponse":
      "properties":
        "date":
          "type": "string"
          "format": "date"
          "title": "Date"
        "items":
          "items":
            "oneOf":
              -
                "$ref": "#/components/schemas/DayEventItem"
              -
                "$ref": "#/components/schemas/DayAssetItem"
            "discriminator":
              "propertyName": "item_type"
              "mapping":
                "asset": "#/components/schemas/DayAssetItem"
                "event": "#/components/schemas/DayEventItem"
          "type": "array"
          "title": "Items"
      "type": "object"
      "required":
        - "date"
        - "items"
      "title": "DayDetailResponse"
      "description": "Serialized response for a single UTC calendar day."
    "DayEventItem":
      "properties":
        "item_type":
          "type": "string"
          "const": "event"
          "title": "Item Type"
        "id":
          "type": "integer"
          "title": "Id"
        "timestamp":
          "type": "string"
          "format": "date-time"
          "title": "Timestamp"
        "event_type":
          "type": "string"
          "title": "Event Type"
        "title":
          "type": "string"
          "title": "Title"
        "summary":
          "anyOf":
            -
              "type": "string"
            -
              "type": "null"
          "title": "Summary"
      "type": "object"
      "required":
        - "item_type"
        - "id"
        - "timestamp"
        - "event_type"
        - "title"
        - "summary"
      "title": "DayEventItem"
      "description": "Serialized day-detail item for a canonical event."
    "ExplorationBootstrapResponse":
      "properties":
        "range":
          "$ref": "#/components/schemas/ExplorationRange"
        "view_modes":
          "items":
            "$ref": "#/components/schemas/ExplorationViewMode"
          "type": "array"
          "title": "View Modes"
        "persons":
          "items":
            "$ref": "#/components/schemas/ExplorationPerson"
          "type": "array"
          "title": "Persons"
        "tags":
          "items":
            "$ref": "#/components/schemas/ExplorationTag"
          "type": "array"
          "title": "Tags"
      "type": "object"
      "required":
        - "range"
        - "view_modes"
        - "persons"
        - "tags"
      "title": "ExplorationBootstrapResponse"
      "description": "Lightweight shell metadata for exploration initialization."
...
    "ExplorationPerson":
      "properties":
        "id":
          "type": "integer"
          "title": "Id"
        "name":
          "type": "string"
          "title": "Name"
        "role":
          "anyOf":
            -
              "type": "string"
            -
              "type": "null"
          "title": "Role"
      "type": "object"
      "required":
        - "id"
        - "name"
        - "role"
      "title": "ExplorationPerson"
      "description": "Person item used by the exploration shell."
    "ExplorationRange":
      "properties":
        "start":
          "type": "string"
          "format": "date"
          "title": "Start"
        "end":
          "type": "string"
          "format": "date"
          "title": "End"
      "type": "object"
      "required":
        - "start"
        - "end"
      "title": "ExplorationRange"
      "description": "Resolved inclusive date range for exploration responses."
    "ExplorationTag":
      "properties":
        "path":
          "type": "string"
          "title": "Path"
        "label":
          "type": "string"
          "title": "Label"
      "type": "object"
      "required":
        - "path"
        - "label"
      "title": "ExplorationTag"
      "description": "Tag item used by the exploration shell."
    "ExplorationViewMode":
      "properties":
        "id":
          "type": "string"
          "title": "Id"
        "label":
          "type": "string"
          "title": "Label"
        "description":
          "type": "string"
          "title": "Description"
      "type": "object"
      "required":
        - "id"
        - "label"
        - "description"
      "title": "ExplorationViewMode"
      "description": "Backend-defined exploration view mode metadata."
...